refpolicy-mcs: Only relabel rootfs when installing package on host for all machine types.
We detect the case where the package is being installed in the build
by checking for $D. This is done in the generic pkg_postinstall function
to cover all machines using the refpolicy.

The xenclient_dom0_append function addes setfiles for /config and /storage and we do a similar
check for $D here as well.

Conflicts:
	xenclient/recipes/selinux/refpolicy-mcs_2.20130424.bbappend