pseudo-xattr-support: remove invalid UNIQUE constraint
This addresses an issue where an incremental rebuild of the dom0 or
ndvm images will create an image with incorrect/missing selinux labels.

When rebuilding an image, the method used in the pseudo-xattr-support
patch tracks via dev+inode+xattrname.  But there is a UNIQUE constraint
"UNIQUE (path, xattrname)", which is invalid when a file is overwritten
because the inode changes. The other constraint,
"UNIQUE (ino, dev, xattrname)" is sufficient.

OXT-6

Signed-off-by: Chris Patterson <pattersonc@ainfosec.com>