Use ACPI guest firmware that matches upstream QEMU
We used to use the ACPI firmware for QEMU traditional but now we want the
version that matches our new QEMU. Note gpe0_blk_len = ACPI_GPE0_BLK_LEN_V1
is left intact. This matches the new QEMU and is 1/2 as big as the traditional
GPE space. This is also why all of the GPE's in the patch are changed to fit
in 16 bits now. Finally we let the Seabios ACPI builder use the new DSDT
version by removing where we patched it to use traditional.

OXT-259

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>