firstboot: update passwd file and fixup owner of tss files
For the upgrade case, care needs to be taken to ensure the passwd
file matches up with the newly shipped dom0 image.  In doing so,
we need to make sure the files outside of that (/boot, /config,
/storage, etc.) have updated perms.

A quick survey shows most files and dirs are owned root:root,
with the notable exception of /boot/system/tpm.  So we chown
these files.  Other files may be discovered and need to be
handled, but given the limited scope of the files in question,
this makes sense given the complexity of a more generalized
solution to deal with the problem (for now).

OXT-272

Signed-off-by: Chris Patterson <pattersonc@ainfosec.com>