Allow custom source cert filenames
The previous recipe requires the source filenames to be
"prod-cacert.pem" and "dev-cacert.pem" otherwise do_install would fail
to find what it was expecting. This patch modifies do_install to use the
basename from the configuration file for these certificates and put them
into the package in their standardized format without requiring an
intermediate standard that was also different.

Signed-off-by: Kevin Pearson <kevin.pearson.4@us.af.mil>