[pvusb] Add usbback support for URB cancelation.
The Linux usbfront driver must support an URB dequeue callback. When
invoked it is supposed to kill the URB in question if it can and complete
it. To do this an internal command is sent to usbback where it uses the anchor
list to find the URB and kill it if it is still being processed.

OXT-38

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>