Allow QEMU to be notified when guests wake from S3.
OXT-440

Prior to this patch, QEMU requires a wake signal to be communicated
from the input server via the DMbus, but:
  - This is nonideal, as it requires the input server to be running, and
    requires the input server to circumvent the toolstack for waking guests.
  - This means that guests are not woken properly when using toolstack calls,
    such as "xec-vm -n <domain-name> resume".

This commit modifies QEMU to also respond to wake requests via a XenStore node.
Ideally, this information would be communicated via a protocol like QMP; this is
to be considered a stand-in until QMP support is integrated.

NOTE:
While the added patch could be merged into acpi.patch, in this case it seems to
make more sense to leave it separate. Ideally, this patch will eventually be
succeeded by a QMP-based solution, and this patch can just be removed without
having to extract anything from acpi.path.

Signed-off-by: Kyle Temkin <temkink@ainfosec.com>