[linux] Amend patch documentation
As the override interface went away, we now have our own calls to
set/reset the p2m as we like. There is no need to maintain a _legacy
interface since this is the only object to use that trick. As changes
were done, the comment in the patch header was left mentioning the
defunct _legacy calls.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>