Move labeling of /boot/system to selinux-load.sh.
There is no reason to do it prior to the switch_root, and
evidently this was not working correctly.  At the least, it would
have required passing -r /root to setfiles for an alternate root,
and even then it did not appear to work for some reason.  Just take
it to selinux-load.sh, where we can do it after loading policy
and use restorecon instead.

OXT-463

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>