[linux-4.1] fix i915_gem_foreign memory corruption
Although this has only been seen with OpenXT/Linux 4.3 and later, this
has been reproduced using earlier non-stable versions.

Upon release, the GEM object could end up corrupting the slab meta-data
or triggering a double free bug.
Refactor the object to use i915 GEM allocation helpers so GEM can safely
call kmem_cache_free() on the allocated object. This mostly means moving
the fields previously allocated right after the drm_i915_gem_object to
the structure itself, or rather in the union field that seems
appropriate for that (it is used for userptr as well).

This is a good opportunity to do some cleanup & coherency modifications
in naming and such, implying not change to the code behaviour.

Also remove the gem-foreign-debug.patch and instead use DRM_DEBUG macros
where they would be useful to reduce patch-queue overhead and for
general cleanliness.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-478