Fix labeling of files/directories created on firstboot.
/boot/system/tpm is created via install -d from init.root-ro.
coreutils install is supposed to look up the proper security context
and assign it but there seems to be a regression for directory creation,
leaving it in the parent directory type.

Several /config files are created by firstboot.sh via shell redirection.
These will default to the type of the parent directory without an explicit
label assignment.

Add restore calls as appropriate (restore is defined as a shell
wrapper for restorecon in /etc/init.d/functions).
The alternative would be to use name-based type transitions in policy,
but explicit restorecon seems more appropriate here given that these
are being created manually by an init script and are essentially one-time
operations.

OXT-486

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>