libselinux: Apply two upstream fixes required for installer.
selinux upstream commits 9df4988 and 5a8d8c4.
Otherwise, when switching /etc/selinux/config to enforcing for dom0,
the installer fails when attempting to mount selinuxfs, with:
    Mount failed for selinuxfs on /sys/fs/selinux:  No such file or directory

The scenario is that the installer is booted with selinux=0, but
has the same /etc/selinux/config as dom0, and thus has SELINUX=enforcing
if we make dom0 enforcing.  The expected behavior is that kernel
parameters take precedence (override) the /etc/selinux/config setting,
so this used to work.  When the preferred selinuxfs mount point moved from
/selinux to /sys/fs/selinux (in Linux 3.0), attempting to mount selinuxfs
on a kernel with SELinux disabled (via selinux=0 or via kernel config)
started returning ENOENT rather than ENODEV because the /sys/fs/selinux
mount point is only created by the kernel if SELinux is enabled.  This
then confused libselinux into thinking that the mount failed for some
reason other than SELinux being disabled in the kernel, and caused it
to fail-closed, thereby causing init to exit and the system to panic.
An early fix for this was b3b19fdce, which checked /proc/filesystems
to see if SELinux was disabled (i.e. no selinuxfs entry).  But this
fails if init has not already mounted /proc, which seems to be the case
for sysvinit.

Commit 9df4988 fixed this issue by mounting /proc before reading /proc
filesystems, but triggers unnecessary attempts to mount /proc for other
users of libselinux besides init.  Commit 5a8d8c4 fixed that issue by
making the mounting of /proc conditional on it not already being mounted.

Arguably if the installer is going to run with SELinux disabled, then
it shouldn't have any /etc/selinux/config or other SELinux policy files
present, but that is left to a separate change.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>