[monit] Remove
While monit could be useful in production for error recovery and
resource starvation management, currently it is only restarting daemons
in dom0 on failure which can hide (and already has hidden) issues.

I understand this is debatable, as testing should be done on the daemons
initialy and monit, only being a failsafe for now, could provide more
help on resource based events and the likes.
But, for now, I believe this is not used as such and can get in the way
of development.

OXT-514

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>