Add RDEPENDS for libgcc to trousers.
tcsd depends on libgcc_s.so.1 for pthread_cancel.
The absence of libgcc_s.so.1 from the installer rootfs was
causing tcsd to abort during installation, and thereby
preventing installing with measured launch enabled.
libgcc_s.so.1 was already being included in the dom0 rootfs,
which is why tpm-setup could work.

With this fix and the luksCheckKey fix applied, one can
once again install with measured launch enabled.

OXT-473
OXT-518

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>