Disable unnecessary policy modules.
Disabling unused policy modules is helpful in reducing disk
and kernel memory footprint and ensuring that we are not allowing
unexpected behavior/functionality.  The policy modules disabled
in this change do not appear to correspond to any packages installed
in dom0 or the ndvm.

In order to disable the hal policy module, we also need to wrap
any references to it in other modules with optional_policy blocks
in order to allow the policy to build without it.  This required
updating the policy patch for the init module and a couple of the
OpenXT-added policy modules.  Using optional_policy blocks is
preferred regardless for any module that is not a hard dependency.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>