install-bootloader: restorecon grub-language.conf and boot-images
install-bootloader invokes install-bootloader-images, which
deletes and re-creates the grub-language.conf file and the
/config/boot-images symlink.  Run restorecon on these files after
re-creating them so that they are labeled properly.

language-sync also runs install-bootloader-images and therefore needs
to run restorecon aka setfiles after this change, so adjust its policy
accordingly.

The alternative would be to define a domain for install-bootloader
and define type transitions for these files to automatically
label them differently than the parent directory.  The language-sync
domain already has such type transitions defined, which is why it did not
have this problem.  However, install-bootloader is only ever called
by hand from a newrole'd shell and only for development purposes (and
further requires a setenforce 0 presently to work), so it is unclear if
this warrants a separate domain.

OXT-537

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>