[xenstored] Clean-up the packaging & deployment
xen-tools installs and registers the xenstored initscript, which in the
end calls the xenstored binary shipped with the xenclient-toolstack
package.

As we currently use the forked OCAML implementation of xenstored, have a
split package xenclient-toolstack-xenstored that will install and
register its own xenstored initscript for the OCAML forked xenstored.
Fix the xen-tools xenstored initscript to use the C xenstored shipped
with xen-tools-xenstored.

In the end change the dom0 package-group to install the
xenclient-toolstack OCAML xenstored only.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>