[acpi] Move BAT2 status port to prevent AMLI errors in Windows 10
The original guest battery ACPI code had used port 0x90 as the BAT2 status
register. For years this was causing harmless messages in the system log but
nothing was done about it and it was forgotten. With Windows 10, use of this
port started causing endless AMLI debugger errors to trace out from the
guest. This moves the port to someplace that is not problematic to Windows.

OXT-606

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>