OXT-668: intel microcode update to 20160714
This recipe should be replaced by a dependency on the meta-intel
layer, but unfortunately at the moment the microcode version in
the upstream layer is not from 2016. An upstream patch has been
submitted to correct that and we update the OpenXT version here in
the meantime.

refs: OXT-668

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>

cherry-picked from commit: b6a172f18d2b28d4632b1b9b671607ed8fd44e40