Allow running lsusb from a newrole'd root shell.
Running lsusb from a newrole'd root shell fails with
the following SELinux denials:
    avc:  denied  { create } for  pid=1556 comm="lsusb" scontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tcontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tclass=netlink_kobject_uevent_socket
    avc:  denied  { bind } for  pid=1556 comm="lsusb" scontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tcontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tclass=netlink_kobject_uevent_socket
    avc:  denied  { read } for  pid=1556 comm="lsusb" scontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tcontext=root:sysadm_r:sysadm_t:s0-s0:c0.c1023 tclass=netlink_kobject_uevent_socket

Permit this by allowing sysadm_t to create and use netlink sockets with the
NETLINK_KOBJECT_UEVENT protocol.  The alternative would be to introduce
a separate domain for lsusb or other programs that require such access,
but this doesn't seem justified given that we only want to allow this from
a newrole'd root shell and this additional access is not significant for
sysadm_t.

OXT-677

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
(cherry picked from commit 5bf09e704a72ddafed006a8130ed9345d9629b44)