STABLE-5: refpolicy: fix OTA-related denials
Allow cryptsetup to read /run/udev/queue.bin.

Allow dmsetup and mount to statfs sysfs and proc filesystems.

Allow updatemgr to statfs ext4 filesystems.

Allow updatemgr to read /config/repo-cert.conf.
This should be allowed already but several /config files
are mislabeled in 5.x installs.  We aren't likely to fix that
now so we might as well just allow it here for OTA purposes.
Remember to restorecon these files as part of a migration script
or on first boot somewhere.

Allow updatemgr to statfs / stat selinuxfs.

Allow updatemgr to manage the /storage/update/upgrade-data directory;
this was supposed to already be allowed but the wrong type was used in
the xc-installer.if and xc-installer.fc files.

Allow updatemgr to convey environment variables, signal
masks, and resource limits to the installer across a domain transition.

OXT-734

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>