refpolicy: split policy for rpc-proxy vs dbusbouncer
rpc-proxy and dbusbouncer are two different programs, one of which
runs in dom0 and the other in the ndvm, and they have different
interactions with other processes in the system, but they were
running in the same domain, dbusbouncer_t.  Split rpc-proxy into
its own rpcproxy_t domain and specialize both domains to only allow
what is required for each program.  Delete legacy types and rules
that are not in use in OpenXT.

OXT-799

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>