[tboot] Reserve TPM mmio region 0xfed40000
This region is only reported in one of the SSDT, not the e820. On a
system using a 32bit kernel and enough memory, the kernel might end up
making this region a RAM buffer, disabling the TPM.

devm_ioremap() was used (prior 4.6?), avoiding the
devm_request_mem_region() that currently fails under the mentioned
circumstances.

In this case, since TBoot has all the right tool to check for TPM and
add that region, use that to change the e820 and mark
[0xfed4000-0xfed44fff] reserved so the kernel won't use it as RAM
Buffer.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-888