[linux] Force devm_ioremap on TPM IO resource.
This region is only reported in one of the SSDT, not the e820. On a
system using a 32bit kernel and enough memory, the kernel might end up
making this region a RAM buffer, disabling the TPM.

devm_ioremap() was used (prior 4.6?), avoiding the
devm_request_mem_region() that currently fails under the mentioned
circumstances.

In this case revert to use devm_ioremap() anyway as the TPM module used
to.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-888