Allow writable CDs and fix guest crash on write.
This fixes CD writing to what it is in stable-6 currently. Some things fail
like formatting the CD or dragging files/burning. Other things work like
right clicking an ISO and burning it to a CD. A separate ticket will be
opened to track those failures but this fixes the RO problem and the crash.

OXT-931

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>