qemu-dm: Disable devices in generated config
The original approach disabled the devices in the default configuration
files.  While it worked, it meant you could not readily re-enable a
previously disabled device.  A clean set of source files (like from
re-running after 'cleansstate') was needed to re-enable a device.

The new approach generates the target-specific config, modified it, and
then continues the rest of the build.  As the default configs are left
as-is, devices can readily be re-enabled.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>