tboot: adjust TPM eventlog location
For certain devices the memory map resulted in the log landing in the
e820 ram section adjucent to tboot. This caused issues as tboot aligns
all mbi modules to the end of tboot reserved section without regard of
e820 regions. This moves the log to the end of the highest available ram
region which should ensure no conflict.

OXT-1004

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>