[Xen 4.9] Update libxl-xenmgr-support.patch
Rebased for Xen 4.9.0.
Includes a fix to xenstore notification of creating-devices.

The order of operations in domain creation in libxl has changed
since Xen 4.6. XenMgr listens for this xenstore event in order
to synchronize domain creation work.

This change restores the xenstore write to the required point.

OXT-1203 for Xen upgrade.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>