tpm-functions: Improve TPM2 active check
Some BIOS allow the TPM to be "on" but disabled.  In that case we need
to verify actual usability.  Verify that the Platform, Storage,
Endorsement, and Platform NV hierarchies are all enabled to determine if
the TPM2 is active and usable.

At least for a Dell BIOS, a TPM2 that is "on" but "disabled" will have
shEnable and ehEnable set to "clear" in `tpm2_dump_capabilities
--capability=properties-variable`.

OXT-1212

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>