webkit: Upgrade to 2.12.5.
This is the version shipped with morty.

libwebkitgtk is growing quite big, so big that it seems the i686 native
environment does not handle largefile correctly and corrupts the binary
in the end. Building the 2.14.x version shipped with pyro is not easy.
In this version the library is "only" 2.0G, which is already a lot.
Something should be done to shrink it (feature turn-off etc).

This is quite a problem.

Downgrading should be seen as a work-around for now. Fixing the
large-file support would be ideal, upgrading to a 64 bits native->target
combo would be great too.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>