xsession: Refresh and update xinitrc.
UIVM is meant to be a kiosk with surf in full screen and nm-applet in
the up-right corner, started as root.
xfce4-session is not used (that could be done later), leaving xinitrc
the job to start each and every required app:
- keymap-sync
- xfwm4
- resized
- xfsettingsd --no-daemon --replace
  * See later why --no-daemon...
- surf
- language-sync
- nm-backend-sync
- nm-applet-launcher
- uim-toolbar-gtk-wrapper

If any of these end, terminate all child processes (pkill -P $$) to end
the graphic session.

This will restart the whole environment as /etc/inittab maintains
startxfce4 as respawn.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>