xen-ocaml-libs: Unbreak sysroot separation.
As documented in:
https://caml.inria.fr/pub/docs/manual-ocaml/native.html
> If -cclib or -ccopt options are passed on the command line, these
> options are stored in the resulting .cmxalibrary.

So _do not_ override LDFLAGS. Instead use CC directly to define the
sysroot. This will otherwise create library objects that will try to
link with the recipe-sysroots of the providing recipe (so, in its own
workdir).  Even worse, ocamlfind will end up specifically
configuring ocamlopt to link using that sysroot! Usual symptom will be
ld: cannot find -l<stuff> although <libstuff> is in the sysroot of the
recipe you're trying to build.

Maybe this could be done in a more "general" manner using -noautolink
when using findlib, but then the linker will have to figure out all the
-cclib to pass down.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>