policycoreutils: Update audit and pam support
See upstream commit:
https://github.com/SELinuxProject/selinux.git
89ce96ca policycoreutils: make audit and pam support configurable

OE note:
newrole will segfault if empty-root-password is given to the image and
PAM is not in use:
authenticate_via_shadow_passwd() does not check the return value of
crypt(), which will fail and return NULL if the salt is not correct
(empty), which is where OE image with empty-root-password end up.

Quoting shadow(5) manual:
> This field may be empty, in which case no passwords are required to
> authenticate as the specified login name. However, some applications
> which read the /etc/shadow file may decide not to permit any access at
> all if the password field is empty.
So this should probably be fixed and submitted upstream.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>