init.sh: Read runlevel from kernel command line
If we want init to run a runlevel different than the default, it must be
passed to init when started.  Follow convention by reading '0-6', 's'
and 'S' from the kernel command line as the runlevel.

The runlevel is passed to selinux-load.sh which then calls /sbin/init.root-ro
for a measured launch or /sbin/init directly.  /sbin/init.root-ro will call
/sbin/init passing along the runlevel as well.

OXT-1300

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>