dbus: Remove argo.patch
With the converison of syncvm to use INET_IS_ARGO=1, there are no users
of family=argo for DBUS_SYSTEM_BUS_ADDRESS.  Hence, we can just drop the
patch.

The argo patch itself was a little strange.  It allowed family=argo to
set PF_XENARGO, but it relied on LD_PRELOAD of libargo to actually use
argo.

OXT-1306
OXT-1732

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>