dbus: Fix recipe DEPENDS statement.
DEPENDS applies for the whole recipe, giving it an "_<PACKAGE>" will
have Bitbake ignore it.

DBus does not require xen as depend (libv4v does).
Since libv4v is optional for that recipe, depend on it only for the
target build. DBus recipe provides native tools, and libv4v is not
needed for that.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1306