xen: Ammend PGC_pinned_by_tools declaration.
PGC_pinned_by_tools is a flag in page->count_info field, hence
PGC_count_width should be incremented to accomodate for it. Ref-count
has to be quite high to make this fail, which is why no misbehavior was
seen until now.

XenServer maintains an equivalent hunk in one patch:
https://github.com/xenserver/xen-4.7.pg/blob/1b778262c1b4a3c7c8e504d50e0149fe5909d670/master/nvidia-hypercalls.patch#L487

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1258