Revert "v4v: Fix SMAP & HVM guest use"
The original change didn't fix v4v for HVMs with SMAP on Xen 4.9.  It
relied on smap_check_policy in guest_walk_tables to skip the SMAP check,
but that was removed between Xen 4.6 and 4.9.  Upstream NACK-ed
re-introducing support in guest_walk_tables and suggested wrapping the
v4v hypercall in the guest kernel with stac/clac.  We'll do that in the
v4v module.

This reverts commit 816157744a05a32e206ddf2eddd504bc0cb09587.

OXT-1304

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>