rsyslog: fix segfault after configuration errors
Backport upstream fix:
https://github.com/openembedded/meta-openembedded/commit/7b7a49b1bae9204e7c8997e92ab5ccb076df9bc0#diff-634054a5b86172511e9c2c3a8c2877d5

rsyslog will segfault on startup if
a) the local machine's hostname is set to a non-FQDN name
b) the getaddrinfo() system call fails
This scenario is higly unlikely, but may exist especially with
provisioned VMs which may not properly be able to do name queries
on startup (seen for example on AWS).

This patch fixes the situation and also provides more robustness
for very early startup error messages when some of the error-reporting
subsystem is not yet properly initialized. Note that under these
circumstances, errors may only show up on stderr.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1326