xen-libxl: Remove unnecessary code change
libxl-xenmgr-support changed '-M' QEMU flag to '-machine'.  But the two
options are just the short and long options of the same thing.  So there
is no need for the change and it can be dropped.

libxl-move-extra-qemu-args-to-the-end.patch needs fixup for the previous
change.

This keeps libxl closer to upstream.

OXT-1327

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>