usbback: Avoid BUG() on memory exhaustion.
usbback implements two requests types to communicate between front and
back: direct and indirect requests. Direct requests can pass up to 17
grefs while indirect are up to 1023 grefs (used by some bulk/isochronous
devices afaict).

usbback will allocate a given amount of pages, using the modparam "reqs"
which is a 17 pages multiple, 17 being the maximum number of pages a
non-indirect request can pass. Default is 64 reqs, so 1088 pages.

Indirect and direct requests will use that same pool of pages, that
means it is not too difficult to exhaust the pool. When that happens:
-> usbif_schedule
  -> do_usb_io_op
    -> dispatch_usb_io
      -> populate_req
      * fails, goto fail_response
      -> free_req
        -> BUG(!segment)

The current implementation will BUG() if free_req is called on a
request that has segments not populate, which is what happens when the
64 * 17 pool of pages is exausted, especially when indirect requests
used more pages from the pool.

Tracking the populated segments to release only these would be safer.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1319