qemu-dm: Avoid segfault in dmbus_send
acpi.patch adds this call chain:
acpi_pm1_cnt_write
  xenstore_update_power
    xen_input_send_shutdown
      dmbus_send

When xen_input has not been initialized, dmbus_send segfaults since
input.service is NULL.  If we aren't using xen-input, then we don't need
to notify input_server.  That way we avoid the segfault.

Maybe this should be moved to xen_input_deinit?  In that case, we
wouldn't have the shutdown reason on hand - it would need to be stashed
somewhere.

OXT-1360

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>