qemu-dm: Import qemu-dm-wrapper from dm-agent
qemu-dm-wrapper sets up some alsa audio configuration before exec-ing
qemu.  svirt-interposer to launch qemu, but it is packaged in dm-agent.

We are removing dm-agent, but we still need the file, so move it to this
recipe.  svirt-interposer really needs the file, but it should be used
outside svirt-interposer for corrected alsa configuration.

OXT-1381

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>