linux: CONFIG_BLK_DEV_TAP=y in openxt-installer
Set CONFIG_BLK_DEV_TAP=y so the blktap tap-ctl command works in the
installer again.  This was previously disabled since blktap breaks
running as a 32bit compat userspace on a 64bit kernel.  In the meantime,
openxt-installer was converted to a 64bit userspace, so the
compatibility issue is no longer present.

OXT-1378

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>