xen: Bypass ACPI SLIC support for PVH
acpi-slic-support.patch changes acpi_build_tables to modify the OEM ID
so a passed through SLIC table works for Windows activation.  It does so
by modifying the ACPI tables on the fly.  This works fine when run as
part of hvmloader, but it segfaults when the code runs in the xl process
for a PVH domain.

The xl segfault is in is_slic() trying to accessing memory in the range
0xfc00xxxx, ACPI_INFO_PHYSICAL_ADDRESS.  acpi_build_tables calls v2p to
convert pointers to the destination physical address.  In hvmloader
this is a no-op, so continued access through those pointers works.  With
xl for a PVH domain, it does an actual conversion to the destination
ACPI memory range which is invalid for the userspace process.

A SLIC entry can only be present if it is passed in, and it will never
be passed in for a PVH.  So conditionalize execution on whether a table
is passed in or not to avoid the fault.

While here, switch to using nr_secondaries to limit access to only the
populated entries of secondary_tables.

OXT-1383

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>