refpolicy-mcs: Allow loading signed modules
When loading a signed module, the kernel keyring must be searched to
find the certificate for verification.  This is an independent
permission and without it you'll see:
"PKCS#7 signature not signed with a trusted key"

For OpenXT, udev will load modules with its built-in kmod support.
However udev_t does not have permission to search the keyring -
resulting in the above message.  Instead of adding the permission
whack-a-mole to individual types, add it to the kernel_load_module
interface since it will always be needed.

OXT-1403

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>