module-signing: Allow external keys
Allow specifying of keys to the build system.  The default when
unspecified is to create and sign with an auto-generated keypair.

For external keys, two variables can be set in local.conf:
KERNEL_MODULE_SIG_CERT and KERNEL_MODULE_SIG_KEY.
KERNEL_MODULE_SIG_CERT can point to the absolute filepath of a public
key for signed module verification.
KERNEL_MODULE_SIG_KEY can point to the absolute filepath of a private key
for signing kernel modules.

If KERNEL_MODULE_SIG_CERT is defined, it will be embedded in the
generated linux kernel.
If KERNEL_MODULE_SIG_KEY is defined, it will be used to sign the modules
during the build.  If undefined, modules must be signed manually.
Images would need to be extracted, all the .ko files signed in
/lib/modules, and then the image re-packed.

Keypairs can be generated according to
Documentation/admin-guide/module-signing.rst in the linux source tree.

OXT-1403

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>