openxt-ml: Do not mount ESP in seal-system.
seal-system is called from the installer install-main:seal_system() only
on upgrade. It is executed as chroot ${DOM0_MOUNT} /usr/sbin/seal-system
-f -r ${ROOT_DEV}. seal-system is mounting ESP to grab all the files
needed for forward sealing. These were copied from the new dom0 image on
$DOM0_MOUNT into ESP in the first place copy_to_esp.

Knowning that, assume default path for the measured files and provide
two new options to seal-system:
-4  [object[,objects]]  List, comma ',' separated, of objects measured
                        in pcr4 for UEFI installations. The hash of the
                        first object in the list is calculated by
                        pesign, while the following will extend the
                        Autenticode hash, measured as raw-buffer.
-8  [object[,objects]]  List, comma ',' separated, of objects measured
                        in pcr8 for UEFI installations. Order is
                        significant.

From PR discussion:
Note that the unusual behavior of `-4' is a work-around. shimx64.efi
will be loaded by the firmware using Authenticode hash according to the
specification, requiring it to be measured with pesign. Due to a
firmware bug on Dell systems, the other executables can't be measured
with Authenticode and instead are measured as raw buffers.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1397